Split Redis coordinator from Redis runner.
They are two separate things, united only in using Redis as a backend.
