+++ funflow.cabal
@@ -40,6 +40,7 @@ Library
                    , Control.FunFlow.External.Executor
                    , Control.FunFlow.External.Coordinator
                    , Control.FunFlow.External.Coordinator.Memory
+                   , Control.FunFlow.External.Coordinator.Redis
                    , Control.FunFlow.Steps
                    , Control.FunFlow.Pretty
                    , Control.FunFlow.Exec.Simple
